requirements.txt
Django==3.1.5
djangorestframework==3.12.2
psycopg2==2.8.6
Pillow==8.0.1